requirements.txt
hydra-core>=1.2
numpy>=1.24
polars>=0.20
pyarrow>=14
scipy>=1.10
lightgbm>=4.0
optuna>=3.0
ray[tune]>=2.9
matplotlib>=3.8
pillow>=10.0
